refactor: move db/types.ts to types/item.ts
- Moved src/db/types.ts to src/types/item.ts for consistent organization
- Updated all imports in 36 files (scrapers, utils, etc.)
- All types now properly located in src/types/ directory
- No functional changes, just better file organization
