+++ src/scrapers/cos.com.ts
@@ -1,5 +1,5 @@
 import type { Page } from 'playwright';
-import type { Item, Image, Size } from '../db/types.js';
+import type { Item, Image, Size } from '../types/item.js';
 import { uploadImagesToS3AndAddUrls } from '../lib/image-utils.js';
 import { formatItem } from '../db/db-utils.js';
 import { logger } from '../lib/logger.js';
