+++ src/scrapers/griebags.com.ts
@@ -1,7 +1,7 @@
 import { /* chromium, */ Page } from 'playwright';
 // import type { Item, Scraper } from './types.js'; // Corrected import below
 import type { Scraper } from './types.js';
-import type { Item, Image } from '../db/types.js'; // Direct import for Item and Image
+import type { Item, Image } from '../types/item.js'; // Direct import for Item and Image
 import * as Utils from '../db/db-utils.js';
 import { uploadImagesToS3AndAddUrls } from '../lib/image-utils.js';
 import { logger } from '../lib/logger.js';
