+++ src/scrapers/iam-store.com.ts
@@ -1,6 +1,6 @@
 import type { Page, Browser } from 'playwright';
 import { chromium } from 'playwright';
-import { Item, Image, Size, ItemSchema, ImageSchema, SizeSchema } from "../db/types.js";
+import { Item, Image, Size, ItemSchema, ImageSchema, SizeSchema } from '../types/item.js";
 import * as Utils from "../db/db-utils.js";
 import { getSiteConfig } from "../types/site-config.js";
 import { uploadImageUrlToS3 } from '../providers/s3.js';
