+++ src/scrapers/leskizzo.com.ts
@@ -1,5 +1,5 @@
 import type { Page } from 'playwright';
-import { Item, Size, Image } from "../db/types.js"; // Import needed types
+import { Item, Size, Image } from '../types/item.js"; // Import needed types
 import * as Utils from "../db/db-utils.js";
 import type { Scraper } from './types.js';
 import { uploadImageUrlToS3 } from '../providers/s3.js'; // Import S3 function
