+++ src/scrapers/tago.ua.ts
@@ -1,5 +1,5 @@
 import { /* chromium, Browser, */ Page } from 'playwright';
-import { Item, Image, Size } from '../db/types.js';
+import { Item, Image, Size } from '../types/item.js';
 import * as Utils from '../db/db-utils.js';
 import { uploadImageUrlToS3 } from '../providers/s3.js';
 import { uploadImagesToS3AndAddUrls } from '../lib/image-utils.js';
